Class registration: register all default calendars.
