+++ pandas_market_calendars/__init__.py
@@ -14,7 +14,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .market_calendar import MarketCalendar, get_calendar, get_calendar_names
+from .market_calendar import MarketCalendar
+from .calendar_registry import get_calendar, get_calendar_names
 from .calendar_utils import merge_schedules, date_range, convert_freq
 import pkg_resources
