+++ pandas_market_calendars/market_calendar.py
@@ -307,23 +307,6 @@ class MarketCalendar(six.with_metaclass(RegisteryMeta)):
             end,
         )
 
-def get_calendar(name, open_time=None, close_time=None):
-    """
-    Retrieves an instance of an MarketCalendar whose name is given.
-
-    :param name: The name of the MarketCalendar to be retrieved.
-    :param open_time: Market open time override as datetime.time object. If None then default is used.
-    :param close_time: Market close time override as datetime.time object. If None then default is used.
-    :return: MarketCalendar of the desired calendar.
-    """
-    return MarketCalendar.factory(name, open_time=open_time, close_time=close_time)
-
-def get_calendar_names():
-    """All Market Calendar names and aliases that can be used in "factory"
-    :return: list(str)
-    """
-    return MarketCalendar.calendar_names()
-
 def days_at_time(days, t, tz, day_offset=0):
     """
     Create an index of days at time ``t``, interpreted in timezone ``tz``. The returned index is localized to UTC.
